Move BaseUnit to its own file
