+++ core/src/types/mod.rs
@@ -2,16 +2,18 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+mod base_unit;
 mod bigint;
 mod bigrat;
 mod date;
 mod number;
 mod numeric;
 
+pub use base_unit::BaseUnit;
 pub use bigint::{BigInt, BigIntError};
 pub use bigrat::BigRat;
 pub use date::GenericDateTime;
-pub use number::{BaseUnit, Dimensionality, Number, NumberParts, NumberPartsFmt};
+pub use number::{Dimensionality, Number, NumberParts, NumberPartsFmt};
 pub use numeric::{Digits, Numeric, NumericParts};
 
 #[deprecated(since = "0.7.0", note = "renamed to BaseUnit")]
